# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SetApiKeyBody {
  apiKey: string;
}

export interface ApiKeyStatus {
  hasKey: boolean;
  valid: boolean;
  message?: string;
}

export type ChatBodyContext = {
  cwd?: string;
  projectFiles?: string[];
  [key: string]: unknown;
};

export interface ChatBody {
  message: string;
  sessionId?: string;
  dryRun?: boolean;
  context?: ChatBodyContext;
}

export interface SuggestedCommand {
  command: string;
  explanation: string;
  isDangerous: boolean;
  dangerWarning?: string;
  requiresConfirmation: boolean;
}

export type ChatResponseRequirementsCheck = {
  satisfied?: boolean;
  missing?: string[];
  [key: string]: unknown;
};

export interface ChatResponse {
  intent: string;
  response: string;
  commands: SuggestedCommand[];
  requirementsCheck?: ChatResponseRequirementsCheck;
  historyId?: number;
}

export interface ExecuteBody {
  command: string;
  historyId?: number;
  confirmed: boolean;
}

export interface ExecuteResult {
  success: boolean;
  stdout: string;
  stderr: string;
  exitCode: number;
  durationMs?: number;
}

export type ExplainBodyMode =
  (typeof ExplainBodyMode)[keyof typeof ExplainBodyMode];

export const ExplainBodyMode = {
  brief: "brief",
  detailed: "detailed",
  beginner: "beginner",
} as const;

export interface ExplainBody {
  command: string;
  mode?: ExplainBodyMode;
}

export type ExplainResponseFlagsItem = {
  flag: string;
  description: string;
};

export interface ExplainResponse {
  command: string;
  explanation: string;
  flags: ExplainResponseFlagsItem[];
  isDangerous: boolean;
  dangerWarning?: string;
}

export interface SystemCheckBody {
  /** List of checks to run (node, git, python, docker, etc.) */
  checks: string[];
}

export interface SystemCheckItem {
  name: string;
  installed: boolean;
  version?: string;
  path?: string;
  message?: string;
}

export interface SystemCheckResponse {
  results: SystemCheckItem[];
  cwd: string;
}

export type ContextResponsePackageJson = { [key: string]: unknown };

export interface ContextResponse {
  cwd: string;
  projectFiles: string[];
  projectType?: string;
  packageJson?: ContextResponsePackageJson;
  requirementsTxt?: boolean;
  gitInitialized: boolean;
  nodeVersion?: string;
}

export interface HistoryItem {
  id: number;
  input: string;
  intent?: string;
  command?: string;
  explanation?: string;
  executed: boolean;
  success?: boolean;
  output?: string;
  createdAt: string;
}

export interface HistoryResponse {
  items: HistoryItem[];
  total: number;
}

export interface ClearHistoryResponse {
  deleted: number;
}

export type RequirementsResponseRequirements = { [key: string]: string[] };

export interface RequirementsResponse {
  requirements: RequirementsResponseRequirements;
}

export interface AddRequirementBody {
  taskName: string;
  steps: string[];
}

export type FixErrorBodyContext = { [key: string]: unknown };

export interface FixErrorBody {
  command: string;
  stderr: string;
  exitCode?: number;
  context?: FixErrorBodyContext;
}

export interface FixErrorResponse {
  explanation: string;
  suggestedFix: string;
  commands: SuggestedCommand[];
}

export type GetHistoryParams = {
  limit?: number;
};