# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Set OpenAI API key for the session
 */
export const SetApiKeyBody = zod.object({
  apiKey: zod.string(),
});

export const SetApiKeyResponse = zod.object({
  hasKey: zod.boolean(),
  valid: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Check if API key is set
 */
export const GetApiKeyStatusResponse = zod.object({
  hasKey: zod.boolean(),
  valid: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Send a natural language message and get AI response with command
 */
export const chatBodyDryRunDefault = false;

export const ChatBody = zod.object({
  message: zod.string(),
  sessionId: zod.string().optional(),
  dryRun: zod.boolean().default(chatBodyDryRunDefault),
  context: zod
    .object({
      cwd: zod.string().optional(),
      projectFiles: zod.array(zod.string()).optional(),
    })
    .optional(),
});

export const ChatResponse = zod.object({
  intent: zod.string(),
  response: zod.string(),
  commands: zod.array(
    zod.object({
      command: zod.string(),
      explanation: zod.string(),
      isDangerous: zod.boolean(),
      dangerWarning: zod.string().optional(),
      requiresConfirmation: zod.boolean(),
    }),
  ),
  requirementsCheck: zod
    .object({
      satisfied: zod.boolean().optional(),
      missing: zod.array(zod.string()).optional(),
    })
    .optional(),
  historyId: zod.number().optional(),
});

/**
 * @summary Execute a shell command after user confirmation
 */
export const executeCommandBodyConfirmedDefault = false;

export const ExecuteCommandBody = zod.object({
  command: zod.string(),
  historyId: zod.number().optional(),
  confirmed: zod.boolean().default(executeCommandBodyConfirmedDefault),
});

export const ExecuteCommandResponse = zod.object({
  success: zod.boolean(),
  stdout: zod.string(),
  stderr: zod.string(),
  exitCode: zod.number(),
  durationMs: zod.number().optional(),
});

/**
 * @summary Explain what a shell command does
 */
export const explainCommandBodyModeDefault = `brief`;

export const ExplainCommandBody = zod.object({
  command: zod.string(),
  mode: zod
    .enum(["brief", "detailed", "beginner"])
    .default(explainCommandBodyModeDefault),
});

export const ExplainCommandResponse = zod.object({
  command: zod.string(),
  explanation: zod.string(),
  flags: zod.array(
    zod.object({
      flag: zod.string(),
      description: zod.string(),
    }),
  ),
  isDangerous: zod.boolean(),
  dangerWarning: zod.string().optional(),
});

/**
 * @summary Check system requirements
 */
export const SystemCheckBody = zod.object({
  checks: zod
    .array(zod.string())
    .describe("List of checks to run (node, git, python, docker, etc.)"),
});

export const SystemCheckResponse = zod.object({
  results: zod.array(
    zod.object({
      name: zod.string(),
      installed: zod.boolean(),
      version: zod.string().optional(),
      path: zod.string().optional(),
      message: zod.string().optional(),
    }),
  ),
  cwd: zod.string(),
});

/**
 * @summary Get current working directory and project context
 */
export const GetContextResponse = zod.object({
  cwd: zod.string(),
  projectFiles: zod.array(zod.string()),
  projectType: zod.string().optional(),
  packageJson: zod.record(zod.string(), zod.unknown()).optional(),
  requirementsTxt: zod.boolean().optional(),
  gitInitialized: zod.boolean(),
  nodeVersion: zod.string().optional(),
});

/**
 * @summary Get command history
 */
export const getHistoryQueryLimitDefault = 50;

export const GetHistoryQueryParams = zod.object({
  limit: zod.coerce.number().default(getHistoryQueryLimitDefault),
});

export const GetHistoryResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      input: zod.string(),
      intent: zod.string().optional(),
      command: zod.string().optional(),
      explanation: zod.string().optional(),
      executed: zod.boolean(),
      success: zod.boolean().optional(),
      output: zod.string().optional(),
      createdAt: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Clear command history
 */
export const ClearHistoryResponse = zod.object({
  deleted: zod.number(),
});

/**
 * @summary Get all task requirements definitions
 */
export const GetRequirementsResponse = zod.object({
  requirements: zod.record(zod.string(), zod.array(zod.string())),
});

/**
 * @summary Add or update a task requirement
 */
export const AddRequirementBody = zod.object({
  taskName: zod.string(),
  steps: zod.array(zod.string()),
});

export const AddRequirementResponse = zod.object({
  requirements: zod.record(zod.string(), zod.array(zod.string())),
});

/**
 * @summary Get AI suggestions to fix a command error
 */
export const FixErrorBody = zod.object({
  command: zod.string(),
  stderr: zod.string(),
  exitCode: zod.number().optional(),
  context: zod.record(zod.string(), zod.unknown()).optional(),
});

export const FixErrorResponse = zod.object({
  explanation: zod.string(),
  suggestedFix: zod.string(),
  commands: zod.array(
    zod.object({
      command: zod.string(),
      explanation: zod.string(),
      isDangerous: zod.boolean(),
      dangerWarning: zod.string().optional(),
      requiresConfirmation: zod.boolean(),
    }),
  ),
});