# lib/api-zod/src/generated/types/apiKeyStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKeyStatus {
  hasKey: boolean;
  valid: boolean;
  message?: string;
}