# lib/api-zod/src/generated/types/chatBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatBodyContext } from "./chatBodyContext";

export interface ChatBody {
  message: string;
  sessionId?: string;
  dryRun?: boolean;
  context?: ChatBodyContext;
}