# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatResponseRequirementsCheck } from "./chatResponseRequirementsCheck";
import type { SuggestedCommand } from "./suggestedCommand";

export interface ChatResponse {
  intent: string;
  response: string;
  commands: SuggestedCommand[];
  requirementsCheck?: ChatResponseRequirementsCheck;
  historyId?: number;
}