# lib/api-zod/src/generated/types/contextResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ContextResponsePackageJson } from "./contextResponsePackageJson";

export interface ContextResponse {
  cwd: string;
  projectFiles: string[];
  projectType?: string;
  packageJson?: ContextResponsePackageJson;
  requirementsTxt?: boolean;
  gitInitialized: boolean;
  nodeVersion?: string;
}