# lib/api-zod/src/generated/types/explainResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ExplainResponseFlagsItem } from "./explainResponseFlagsItem";

export interface ExplainResponse {
  command: string;
  explanation: string;
  flags: ExplainResponseFlagsItem[];
  isDangerous: boolean;
  dangerWarning?: string;
}