# lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryItem } from "./historyItem";

export interface HistoryResponse {
  items: HistoryItem[];
  total: number;
}