# lib/api-zod/src/generated/types/requirementsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { RequirementsResponseRequirements } from "./requirementsResponseRequirements";

export interface RequirementsResponse {
  requirements: RequirementsResponseRequirements;
}