# lib/api-zod/src/generated/types/suggestedCommand.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface SuggestedCommand {
  command: string;
  explanation: string;
  isDangerous: boolean;
  dangerWarning?: string;
  requiresConfirmation: boolean;
}