# lib/api-zod/src/generated/types/systemCheckBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Terminal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface SystemCheckBody {
  /** List of checks to run (node, git, python, docker, etc.) */
  checks: string[];
}